add authorization to front and backend

+++ backend/routes/spendings.js
@@ -12,7 +12,8 @@ router.post("",
     value: req.body.value,
     description: req.body.description,
     payer: req.body.payer,
-    date: req.body.date
+    date: req.body.date,
+    creator: req.userData.userId
   });
   spending.save().then(createdSpending => {
     res.status(201).json({
@@ -30,10 +31,15 @@ router.put("/:id",
     value: req.body.value,
     description: req.body.description,
     payer: req.body.payer,
-    date: req.body.date
+    date: req.body.date,
+    creator: req.userData.userId
   })
-  Spending.updateOne({_id: req.params.id }, spending).then(result => {
-    res.status(200).json({ message: "Update successful!" });
+  Spending.updateOne({_id: req.params.id, creator: req.userData.userId }, spending).then(result => {
+    if (result.nModified > 0){
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   })
 })
 
@@ -59,9 +65,12 @@ router.get('/:id', (req, res, next) => {
 router.delete("/:id",
   checkAuth,
   (req, res, next) => {
-  Spending.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Spending deleted!'});
+  Spending.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0){
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
